Reference

Contents

Reference

Contents

Index

Index

BinnedDistributionFit.ExtendPdf

BinnedDistributionFit.SumOfPdfs

BinnedDistributionFit.RooFitNLL

BinnedDistributionFit.RooFitNLL_functor

BinnedDistributionFit.RooFitNLL_functor

BinnedDistributionFit.scalar_eval

BinnedDistributionFit.ExtendPdf — Type

struct ExtendPdf{F, S}
    func::F
    support::S
end

Type representing an extended pdf where normalization does not need to equal to 1. func can be any function that can be evaluated within support, when calculating Maximum Likelihood, func will be numerically normalized over support.

Examples

julia> f(x, ps) = x^ps[1] + x*ps[2]

julia> d = BinnedDistributionFit.ExtendPdf(f, (1, 2))

julia> BinnedDistributionFit.scalar_eval(d, 2.0, [1,2])
6.0

julia> BinnedDistributionFit.vector_eval(d, [2.0, 3.0], [1,2])
2-element Vector{Float64}:
 6.0
 9.0

BinnedDistributionFit.SumOfPdfs — Type

struct SumOfPdfs{V, S}
    pdfs::V
    support::S
end

Type representing a sum of pdfs.

Examples

julia> d1 = ExtendPdf((x, _)->x, (1,3))

julia> d2 = ExtendPdf((x, _)->x^2, (1,3))

julia> sd = d1+d2
SumOfPdfs{Vector{ExtendPdf{F, Tuple{Int64, Int64}} where F}, Tuple{Int64, Int64}}(ExtendPdf{F, Tuple{Int64, Int64}} where F[ExtendPdf{var"#13#14", Tuple{Int64, Int64}}(var"#13#14"(), (1, 3)), ExtendPdf{var"#15#16", Tuple{Int64, Int64}}(var"#15#16"(), (1, 3))], (1, 3))

julia> BinnedDistributionFit.scalar_eval(sd, 2.0)
6.0

julia> BinnedDistributionFit.vector_eval(sd, [2.0, 3.0])
2-element Vector{Float64}:
 6.0
12.0

julia> sd2 = d1+d2+d3

julia> BinnedDistributionFit.scalar_eval(sd2, 2.0, [[], [], [1,2]])
ERROR: BoundsError: attempt to access 2-element Vector{Any} at index [3]

julia> BinnedDistributionFit.scalar_eval(sd2, 2.0, [[], [], [1,2,3]])
9.0

BinnedDistributionFit.RooFitNLL — Method

RooFitNLL(normalized_predictions, observations; normalization)

Computes the negative log-likelihood for a binned distribution fit.

normalized_predictions: A vector of normalized prediction at each bin-center.

observations: A vector of observed values at each bin-center.

normalization: A scalar value for normalization. Conceptually equal to the pdf generateing predictions integrated over the domain.

BinnedDistributionFit.RooFitNLL_functor — Method

RooFitNLL_functor(d::ExtendPdf, data_hist::Hist1D; num_integrator=SimpleSumIntegrator())

Given an pdf and a data histogram, return a callable function that evaluates to the negative log likelihood with respect to the data histogram within the d.support domain.

The callable function should be called with N+1 parameter, the +1 being the first argument represening the overall # of events. For example, if user-defined function is f(x, p) = x*p[1] + x^2*p[2], then the functor should be called with [norm, p1, p2].

Examples

julia> f(x,p) = x + p[1] + p[2]*x

julia> d = ExtendPdf(f, (1,3))

julia> BinnedDistributionFit.scalar_eval(d, 1, [2,3])
6

julia> NF = RooFitNLL_functor(d, h);

# the 10. is overall normalization
# [2.0, 3.0] are passed as `p` to `f(x,p)`
julia> NF([10., 2.0, 3.0])
0.060373400847997694

BinnedDistributionFit.RooFitNLL_functor — Method

RooFitNLL_functor(sumpdfs::SumOfPdfs, data_hist::Hist1D; num_integrator=SimpleSumIntegrator())

Given an SumOfPdfs and a data histogram, return a callable function that evaluates to the negative log likelihood with respect to the data histogram within the d.support domain.

The callable function should be called with a vector of vector as input, specifically [norms, ps1, ps2, ...], where norms is the normalization for each of the pdfs in order, and ps1, ps2, ... are the parameters for each pdf function, also in order.

Examples

julia> f(x,ps) = x + ps[1] + ps[2]*x
julia> g(x,ps) = ps[1] + ps[3]

julia> d1 = ExtendPdf(f, (1,3))

julia> d2 = ExtendPdf(g, (1,3))

julia> sd = d1+d2

julia> h = Hist1D(; binedges=1:3, bincounts=[2.0, 4.0], sumw2=[2.0, 4.0])
edges: [1.0, 2.0, 3.0]
bin counts: [2.0, 4.0]
total count: 6.0

julia> NF = RooFitNLL_functor(sd, h; num_integrator=BinnedDistributionFit.QuadGKIntegrator())

julia> NF([[2.0, 3.0], [6.0, 7.0], [1.0, 5.0, 9.0]])
-0.627546320921633

BinnedDistributionFit.scalar_eval — Function

scalar_eval(d::SumOfPdfs, x, vps::Vector{Vector}; kw...)

Evaluate the sum of pdfs at a single point. ps needs to be a vector of vectors of parameters because each pdf in the sum has its own set of parameters.

For example, if we have two pdfs pdf1 and pdf2 with parameters params1 and params2, respectively, we can evaluate the sum of the pdfs at a point x as follows:

params1 = [1.0, 2.0]
params2 = [3.0, 4.0]
scalar_eval(sumofpdfs([pdf1, pdf2]), x, [params1, params2])